requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
python-decouple>=3.8
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
Pillow>=10.1.0
mysqlclient>=2.2.0
gunicorn>=21.2.0
